устранил недостатки
