+++ js/main.js
@@ -1,4 +1,3 @@
-import {getOffers} from './data.js';
-import './filling.js';
+import {renderOffers} from './render.js';
 
-getOffers(10);
+renderOffers(1);
